+++ src/main/java/com/finderfeed/solarcraft/content/blocks/infusing_table_things/InfuserTileEntity.java
@@ -142,12 +142,12 @@ public class InfuserTileEntity extends REItemHandlerBlockEntity implements Solar
     }
 
     @Override
-    public double getMaxRunicEnergyInput() {
+    public float getMaxRunicEnergyInput() {
         return 30;
     }
 
     @Override
-    public double getRunicEnergyLimit() {
+    public float getRunicEnergyLimit() {
         return 100000;
     }
